application/requirements.txt
flask
flask_sqlalchemy
flask_wtf
wtforms
flask_testing
pymysql
pytest
pytest-cov
